support release build for armeabi-v7a; fix some bugs in some CMakeLists.txt; update README.md.

+++ app/src/main/cpp/chromium/src/base/android_build.sh
@@ -1,22 +1,37 @@
-cmake_cmd=/home/dagger/tools/android/sdk/cmake/3.6.4111459/bin/cmake
+#android_sdk_home=/home/dagger/tools/android/sdk
+android_sdk_home=/home/daiguozhou/tools/android/sdk
+
+android_sdk_cmake_version="3.10.2.4988404"
+#android_sdk_cmake_version="3.6.4111459"
+cmake_cmd=${android_sdk_home}/cmake/${android_sdk_cmake_version}/bin/cmake
 #cmake_cmd=cmake
 
-#ndk_root=/home/dagger/tech/code/chromium/src/third_party/android_ndk
-ndk_root=/home/dagger/tools/android/sdk/ndk-bundle
-#ndk_root=/home/dagger/tools/android/ndk/android-ndk-r16b
+if [ ! -f ${cmake_cmd} ]; then
+  echo "ERROR: ${cmake_cmd} does not exist"
+  exit 1
+fi
+
+ndk_root=${android_sdk_home}/ndk-bundle
 #ndk_root=/home/dagger/tools/android/ndk/android-ndk-r17b
+#ndk_root=/home/dagger/tech/code/chromium/src/third_party/android_ndk
+
+android_abi="arm64-v8a" # armeabi-v7a
+android_platform="android-19"
+build_type="Release" # Debug
 
 prj_dir=$(pwd)
-build_dir=${prj_dir}/android_build
+
+
+build_dir=${prj_dir}/android_build/${android_abi}/${build_type}
 
 if [ -d "${build_dir}" ]; then
   rm -fr ${build_dir}
 fi
 
 ${cmake_cmd} -H${prj_dir} -B${build_dir} \
-  -DANDROID_ABI=arm64-v8a \
-  -DANDROID_PLATFORM=android-21 \
-  -DCMAKE_BUILD_TYPE=Release \
+  -DANDROID_ABI=${android_abi} \
+  -DANDROID_PLATFORM=${android_platform} \
+  -DCMAKE_BUILD_TYPE=${build_type} \
   -DANDROID_NDK=${ndk_root} \
   -DCMAKE_TOOLCHAIN_FILE=${ndk_root}/build/cmake/android.toolchain.cmake \
   -DCMAKE_VERBOSE_MAKEFILE:BOOL=ON \
